# internal/platform/server/handlers_object.go
package server

import (
	"github.com/labstack/echo/v4"
	"gitlab.snapp.ir/platform/snapp_object_store/internal/domain/objectstorage"
	"net/http"
)

func HandleObjectDownload(s *Server) echo.HandlerFunc {
	return func(c echo.Context) error {
		var req objectstorage.ObjectRequestMeta
		err := c.Bind(&req)
		if err != nil {
			s.logger.Error(err.Error())
			return c.JSON(http.StatusBadRequest, err.Error())
		}

		err = (&echo.DefaultBinder{}).BindHeaders(c, &req)
		if err != nil {
			s.logger.Error(err.Error())
			return c.JSON(http.StatusBadRequest, err.Error())
		}

		err = c.Validate(req)
		if err != nil {
			return c.JSON(http.StatusBadRequest, err.Error())
		}

		objects, errObjectDownload := s.db.ObjectDownload(s.config.ObjectStorageConfigs, req)
		if errObjectDownload.Message != nil {
			return c.JSON(errObjectDownload.Code, errObjectDownload.Message.Error())
		}
		return c.JSON(http.StatusOK, objects)
	}
}

func HandleObjectUpload(s *Server) echo.HandlerFunc {
	return func(c echo.Context) error {
		var req objectstorage.ObjectRequestMeta
		err := c.Bind(&req)
		if err != nil {
			s.logger.Error(err.Error())
			return c.JSON(http.StatusBadRequest, err.Error())
		}

		err = (&echo.DefaultBinder{}).BindHeaders(c, &req)
		if err != nil {
			s.logger.Error(err.Error())
			return c.JSON(http.StatusBadRequest, err.Error())
		}

		err = c.Validate(req)
		if err != nil {
			return c.JSON(http.StatusBadRequest, err.Error())
		}

		objects, errObjectUpload := s.db.ObjectUpload(s.config.ObjectStorageConfigs, req)
		if err != nil {
			return c.JSON(errObjectUpload.Code, errObjectUpload.Message.Error())
		}
		return c.JSON(http.StatusOK, objects)
	}
}

// HandleObjectList
//
//	@Summary		List of objects of a bucket
//	@Description	Fetches list of buckets owned by a user that is specified by AccessKey and SecretKey
//	@Tags			Object
//	@Accept			json
//	@Produce		json
//	@Param			access_key	header		string								true	"User given AccessKey"
//	@Param			secret_key	header		string								true	"User given SecretKey"
//	@Param			bucket		query		string								true	"bucket name"
//	@Param			max_keys	query		string								true	"max_keys of pagination"
//	@Param			page		query		string								true	"page of pagination"
//	@Success		200			{object}	objectstorage.ObjectListResponse	"Successful response with bucket list"
//	@Failure		400			{object}	string								"Bad Request"
//	@Failure		500			{object}	string								"Internal server error"
//	@Router			/api/object/list [get]
func HandleObjectList(s *Server) echo.HandlerFunc {
	return func(c echo.Context) error {
		var req objectstorage.ObjectListRequestMeta
		err := c.Bind(&req)
		if err != nil {
			s.logger.Error(err.Error())
			return c.JSON(http.StatusBadRequest, err.Error())
		}

		err = (&echo.DefaultBinder{}).BindHeaders(c, &req)
		if err != nil {
			s.logger.Error(err.Error())
			return c.JSON(http.StatusBadRequest, err.Error())
		}

		err = c.Validate(req)
		if err != nil {
			return c.JSON(http.StatusBadRequest, err)
		}

		objects, errObjectList := s.db.ObjectList(s.config.ObjectStorageConfigs, req)
		if err != nil {
			return c.JSON(errObjectList.Code, errObjectList.Message.Error())
		}
		return c.JSON(http.StatusOK, objects)
	}
}

// HandleObjectsDelete
//
//	@Summary		Deletes the list of objects inside a bucket
//	@Description	Deletes a list of objects specified by name
//	@Tags			Object
//	@Accept			json
//	@Produce		json
//	@Param			access_key	header		string								true	"User given AccessKey"
//	@Param			secret_key	header		string								true	"User given SecretKey"
//	@Param			bucket		body		string								true	"bucket name"
//	@Param			objects		body		string								true	"objects names"
//	@Success		200			{object}	objectstorage.ObjectDeleteResponse	"Successful response with objects delete"
//	@Failure		400			{object}	string								"Bad Request"
//	@Failure		401			{object}	string								"Unauthorized"
//	@Failure		500			{object}	string								"Internal server error"
//	@Router			/api/object/list [get]
func HandleObjectsDelete(s *Server) echo.HandlerFunc {
	return func(c echo.Context) error {
		var req objectstorage.ObjectDeleteRequestMeta
		err := c.Bind(&req)
		if err != nil {
			s.logger.Error(err.Error())
			return c.JSON(http.StatusBadRequest, err.Error())
		}

		err = (&echo.DefaultBinder{}).BindHeaders(c, &req)
		if err != nil {
			s.logger.Error(err.Error())
			return c.JSON(http.StatusBadRequest, err.Error())
		}

		err = c.Validate(req)
		if err != nil {
			return c.JSON(http.StatusBadRequest, err.Error())
		}

		objects, errObjectDelete := s.db.ObjectsDelete(s.config.ObjectStorageConfigs, req)
		if err != nil {
			return c.JSON(errObjectDelete.Code, errObjectDelete.Message.Error())
		}
		return c.JSON(http.StatusOK, objects)
	}
}